+++ tooling/cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## \[1.5.3]
+
+### Dependencies
+
+- Upgraded to `tauri-bundler@1.4.2`
+
+## \[1.5.2]
+
+### Dependencies
+
+- Upgraded to `tauri-bundler@1.4.1`
+
 ## \[1.5.1]
 
 ### Bug Fixes