+++ tooling/cli/node/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## \[1.5.3]
+
+### Dependencies
+
+- Upgraded to `tauri-cli@1.5.3`
+
+## \[1.5.2]
+
+### Dependencies
+
+- Upgraded to `tauri-cli@1.5.2`
+
 ## \[1.5.1]
 
 ### Bug Fixes